Reconfigure Database Class
Downloaded database.php from the lab page to replace db.php, as well as
updating folder.php to reference this new file.
